requirements.txt
# VisionAI Eye Hospital - Python Dependencies
# Production-grade eye disease screening application

# Core Flask
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-Login>=0.6.0

# Security
Werkzeug>=3.0.0

# HTTP Requests
requests>=2.31.0

# AI Model - Local Inference (EfficientNetB0 for Eye Disease Detection)
torch>=2.0.0
transformers>=4.30.0
Pillow>=10.0.0

# Google OAuth
Authlib>=1.3.0

# Google Sheets API
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Supabase Database
supabase>=2.0.0

# Environment variables (optional)
python-dotenv>=1.0.0

# Production WSGI server (optional)
gunicorn>=21.0.0
